+++ trunk/game/sprites/StaticSprite.cs
@@ -24,7 +24,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="xPosition">X Position</param>
         /// <param name="yPosition">Y Position</param>
         /// <param name="random">Random number generator</param>
-        public StaticSprite(float xPosition, float yPosition, Random random)
+        public StaticSprite(double xPosition, double yPosition, Random random)
             : base(xPosition, yPosition, random)
         {
             isDestructible = false;
@@ -37,7 +37,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="yPosition">Y Position</param>
         /// <param name="random">Random number generator</param>
         /// <param name="isDestructible">whether sprite can be destroyed (default: false)</param>
-        public StaticSprite(float xPosition, float yPosition, Random random, bool isDestructible)
+        public StaticSprite(double xPosition, double yPosition, Random random, bool isDestructible)
             : base(xPosition, yPosition, random)
         {
             this.isDestructible = isDestructible;
@@ -50,49 +50,49 @@ namespace AbrahmanAdventure.sprites
             return false;
         }
 
-        protected override float BuildJumpingTime()
+        protected override double BuildJumpingTime()
         {
-            return 0f;
+            return 0;
         }
 
-        protected override float BuildWalkingCycleLength()
+        protected override double BuildWalkingCycleLength()
         {
-            return 0f;
+            return 0;
         }
 
-        protected override float BuildWalkingAcceleration()
+        protected override double BuildWalkingAcceleration()
         {
-            return 0f;
+            return 0;
         }
 
-        protected override float BuildMaxWalkingSpeed()
+        protected override double BuildMaxWalkingSpeed()
         {
-            return 0f;
+            return 0;
         }
 
-        protected override float BuildMaxRunningSpeed()
+        protected override double BuildMaxRunningSpeed()
         {
-            return 0f;
+            return 0;
         }
 
-        protected override float BuildStartingJumpAcceleration()
+        protected override double BuildStartingJumpAcceleration()
         {
-            return 0f;
+            return 0;
         }
 
-        protected override float BuildAttackingTime()
+        protected override double BuildAttackingTime()
         {
-            return 0f;
+            return 0;
         }
 
-        protected override float BuildHitTime()
+        protected override double BuildHitTime()
         {
-            return 16f;
+            return 16;
         }
 
-        protected override float BuildMaxFallingSpeed()
+        protected override double BuildMaxFallingSpeed()
         {
-            return float.PositiveInfinity;
+            return double.PositiveInfinity;
         }
         #endregion
