+++ trunk/game/sprites/staticSprites/AnarchyBlockSprite.cs
@@ -44,7 +44,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="xPosition">x position</param>
         /// <param name="yPosition">y position</param>
         /// <param name="random">random number generator</param>
-        public AnarchyBlockSprite(float xPosition, float yPosition, Random random)
+        public AnarchyBlockSprite(double xPosition, double yPosition, Random random)
             : this(xPosition, yPosition, random, BlockContent.Undefined, false)
         {
         }
@@ -57,7 +57,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="random">random number generator</param>
         /// <param name="blockContent">block's content</param>
         /// <param name="isBrick">whether block looks like a breakable brick block (default:false)</param>
-        public AnarchyBlockSprite(float xPosition, float yPosition, Random random, bool isBrick)
+        public AnarchyBlockSprite(double xPosition, double yPosition, Random random, bool isBrick)
             : this(xPosition, yPosition, random, BlockContent.Undefined, isBrick)
         {
         }
@@ -70,7 +70,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="random">random number generator</param>
         /// <param name="blockContent">block's content (default:undefined)</param>
         /// <param name="isBrick">whether block looks like a breakable brick block (default:false)</param>
-        public AnarchyBlockSprite(float xPosition, float yPosition, Random random, BlockContent blockContent, bool isBrick)
+        public AnarchyBlockSprite(double xPosition, double yPosition, Random random, BlockContent blockContent, bool isBrick)
             : base(xPosition, yPosition, random)
         {
             this.isBrick = isBrick;
@@ -125,32 +125,32 @@ namespace AbrahmanAdventure.sprites
             return false;
         }
 
-        protected override float BuildMaxHealth()
+        protected override double BuildMaxHealth()
         {
-            return 1.0f;
+            return 1.0;
         }
 
-        protected override float BuildAttackStrengthCollision()
+        protected override double BuildAttackStrengthCollision()
         {
             return 0;
         }
 
-        protected override float BuildWidth(Random random)
+        protected override double BuildWidth(Random random)
         {
-            return 1.0f;
+            return 1.0;
         }
 
-        protected override float BuildHeight(Random random)
+        protected override double BuildHeight(Random random)
         {
-            return 1.0f;
+            return 1.0;
         }
 
-        protected override float BuildBounciness()
+        protected override double BuildBounciness()
         {
-            return 0.0f;
+            return 0.0;
         }
 
-        public override Surface GetCurrentSurface(out float xOffset, out float yOffset)
+        public override Surface GetCurrentSurface(out double xOffset, out double yOffset)
         {
             blinkCycle.Increment(1);
             bumpCycle.Increment(1);
@@ -158,7 +158,7 @@ namespace AbrahmanAdventure.sprites
             yOffset = 0;
 
             if (bumpCycle.IsFired)
-                yOffset = bumpCycle.CurrentValue / -20.0f;
+                yOffset = bumpCycle.CurrentValue / -20.0;
 
             if (isFinalized)
                 return surface3;
